// ui/src/api/generated/models/image-generation-capability.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ImageGenerationCapability
 */
export interface ImageGenerationCapability {
    /**
     * 
     * @type {Array<string>}
     * @memberof ImageGenerationCapability
     */
    'aspectRatios'?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof ImageGenerationCapability
     */
    'imageToImage'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ImageGenerationCapability
     */
    'maskInput'?: boolean;
    /**
     * 
     * @type {number}
     * @memberof ImageGenerationCapability
     */
    'maxImagesPerCall'?: number;
    /**
     * 
     * @type {Array<string>}
     * @memberof ImageGenerationCapability
     */
    'outputMediaTypes'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof ImageGenerationCapability
     */
    'sizes'?: Array<string>;
    /**
     * 
     * @type {boolean}
     * @memberof ImageGenerationCapability
     */
    'textToImage'?: boolean;
}