// ui/src/api/generated/models/language-capability.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface LanguageCapability
 */
export interface LanguageCapability {
    /**
     * 
     * @type {boolean}
     * @memberof LanguageCapability
     */
    'fileInput'?: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof LanguageCapability
     */
    'imageInput'?: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof LanguageCapability
     */
    'inputMediaTypes'?: Array<string>;
    /**
     * 
     * @type {Array<string>}
     * @memberof LanguageCapability
     */
    'inputSources'?: Array<LanguageCapabilityInputSourcesEnum>;
    /**
     * 
     * @type {boolean}
     * @memberof LanguageCapability
     */
    'reasoningHistory'?: boolean;
}

export const LanguageCapabilityInputSourcesEnum = {
    Data: 'data',
    Url: 'url'
} as const;

export type LanguageCapabilityInputSourcesEnum = typeof LanguageCapabilityInputSourcesEnum[keyof typeof LanguageCapabilityInputSourcesEnum];