// ui/src/api/generated/models/model-capabilities.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.3
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { ImageGenerationCapability } from './image-generation-capability';
// May contain unused imports in some cases
// @ts-ignore
import type { LanguageCapability } from './language-capability';
// May contain unused imports in some cases
// @ts-ignore
import type { ModelCapabilitySources } from './model-capability-sources';

/**
 * 
 * @export
 * @interface ModelCapabilities
 */
export interface ModelCapabilities {
    /**
     * 
     * @type {ImageGenerationCapability}
     * @memberof ModelCapabilities
     */
    'imageGeneration'?: ImageGenerationCapability;
    /**
     * 
     * @type {LanguageCapability}
     * @memberof ModelCapabilities
     */
    'language'?: LanguageCapability;
    /**
     * 
     * @type {ModelCapabilitySources}
     * @memberof ModelCapabilities
     */
    'sources'?: ModelCapabilitySources;
}