// ui/src/api/generated/models/model-capability-sources.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Halo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.25.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ModelCapabilitySources
 */
export interface ModelCapabilitySources {
    /**
     * 
     * @type {string}
     * @memberof ModelCapabilitySources
     */
    'imageGeneration'?: ModelCapabilitySourcesImageGenerationEnum;
    /**
     * 
     * @type {string}
     * @memberof ModelCapabilitySources
     */
    'language'?: ModelCapabilitySourcesLanguageEnum;
}

export const ModelCapabilitySourcesImageGenerationEnum = {
    Remote: 'remote',
    BuiltIn: 'built-in',
    Manual: 'manual',
    Unknown: 'unknown'
} as const;

export type ModelCapabilitySourcesImageGenerationEnum = typeof ModelCapabilitySourcesImageGenerationEnum[keyof typeof ModelCapabilitySourcesImageGenerationEnum];
export const ModelCapabilitySourcesLanguageEnum = {
    Remote: 'remote',
    BuiltIn: 'built-in',
    Manual: 'manual',
    Unknown: 'unknown'
} as const;

export type ModelCapabilitySourcesLanguageEnum = typeof ModelCapabilitySourcesLanguageEnum[keyof typeof ModelCapabilitySourcesLanguageEnum];